tests/wpt/mozilla/tests/mozilla/inline_event_handler.html

<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body onload="t.step(function() { assert_true(event instanceof Event, 'Should be Event'); this.done(); })">
<script>
var t = async_test();
</script>
</body>
</html>
